Code refactor: add extra line and other according to PR comment
