--- src/constants/Utils.js
@@ -1,3 +1,0 @@
-export const HOME = -1;
-export const COMPLETED = 0;
-export const REMAINING = 1;
